third_party/blink/web_tests/virtual/css-anchor-update-enabled/external/wpt/css/css-anchor-position/anchor-scroll-position-try-014-expected.txt

This is a testharness.js-based test.
[FAIL] anchor-scroll-position-try-014 1
  assert_equals: Anchored element should be at the top of anchor expected 94 but got 194
Harness: the test ran to completion.
